perf: put usePrevious custom Hook in _util/ref.ts
